% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{controlInputVariables}
\alias{controlInputVariables}
\title{Check if given variabels are null}
\usage{
controlInputVariables(id, name)
}
\arguments{
\item{id}{String to check}

\item{name}{String to check}
}
\description{
\code{controlInputVariables} checks if given variabels id and name are null.
Returns an error if both variables id and name are null.
Returns warning if both variables id and name are not null.
}
