% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{controlLimit}
\alias{controlLimit}
\title{Check if given limit is smaller than 200}
\usage{
controlLimit(limit)
}
\arguments{
\item{limit}{Integer to check}
}
\description{
\code{controlLimit} checks if  given limit is smaller than 200.
Returns warning if limit is greater than 200 and sets limit to 200.
}
