% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{getAlbum}
\alias{getAlbum}
\title{Search Jamendo database for an album}
\usage{
getAlbum(
  album_id = NULL,
  album_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID")
)
}
\arguments{
\item{album_id}{Album ID to seach for}

\item{album_name}{Album name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about an album.
See \url{https://developer.jamendo.com/v3.0/albums} for more information.
}
\description{
Search Jamendo database for an album. You can search for either an album ID or an album name.
}
\examples{
\donttest{
##Example
 albumID <- getAlbum(album_id = "14866")
 albumName <- getAlbum(album_name = "Deuteranopia Utopia")
}
}
