% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albums.R
\name{getAlbums}
\alias{getAlbums}
\title{Search Jamendo database for several albums}
\usage{
getAlbums(album_ids, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{album_ids}{List of album IDs to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about several albums.
See \url{https://developer.jamendo.com/v3.0/albums} for more information.
}
\description{
Search Jamendo database for several albums
}
