% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artists.R
\name{getArtists}
\alias{getArtists}
\title{Search Jamendo database for several artists}
\usage{
getArtists(artist_ids, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{artist_ids}{List of artist IDs to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about several albums.
See \url{https://developer.jamendo.com/v3.0/artists} for more information.
}
\description{
Search Jamendo database for several artists
}
