% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed.R
\name{getFeed}
\alias{getFeed}
\title{This method returns the editorial feeds that you can find also on the Jamendo homepage.}
\usage{
getFeed(limit = 10, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{limit}{Integer to set the number of displayed search results. Maximum value is 200.}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about feeds from the Jamendo homepage.
See \url{https://developer.jamendo.com/v3.0/feeds} for more information.
}
\description{
This method returns the editorial feeds that you can find also on the Jamendo homepage.
}
