% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{getMyPlaylists}
\alias{getMyPlaylists}
\title{Get your own playlists}
\usage{
getMyPlaylists(token, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{OAuth access token (the previously created OAuth token)}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about your playlists.
See \url{https://developer.jamendo.com/v3.0/playlists} for more information.
}
\description{
Get your own playlists
}
\examples{
\dontrun{
##Example
 token <- jamendoOAuth(app_name="xxxx")
 playlists <- getMyPlaylists(token)
}
}
