% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{getMyUserProfile}
\alias{getMyUserProfile}
\title{Get information about your user profile}
\usage{
getMyUserProfile(token, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about your user profile.
See \url{https://developer.jamendo.com/v3.0/users} for more information.
}
\description{
Get information about your user profile
}
\examples{
\dontrun{
##Example
 token <- jamendoOAuth(app_name="xxxx")
 user <- getUserProfile(token)
}
}
