% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlists.R
\name{getPlaylist}
\alias{getPlaylist}
\title{Search Jamendo database for a playlist}
\usage{
getPlaylist(playlist_id = NULL, playlist_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{playlist_id}{Playlist ID to seach for}

\item{playlist_name}{Playlist name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about a playlist.
See \url{https://developer.jamendo.com/v3.0/playlists} for more information.
}
\description{
Search Jamendo database for a playlist. You can search for either a playlist ID or a playlist name.
}
