% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radios.R
\name{getRadio}
\alias{getRadio}
\title{This method returns information about a jamendo radio}
\usage{
getRadio(radio_id = NULL, radio_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{radio_id}{Radio ID to seach for}

\item{radio_name}{Radio name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about a radio.
See \url{https://developer.jamendo.com/v3.0/radios} for more information.
}
\description{
This method returns information about a jamendo radio. You can search for either a radio ID or a radio name.
}
