% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radios.R
\name{getRadios}
\alias{getRadios}
\title{This method returns the list of existing jamendo radios}
\usage{
getRadios(limit = 10, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{limit}{Integer to set the number of displayed search results. Maximum value is 200.}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about a radio.
See \url{https://developer.jamendo.com/v3.0/radios} for more information.
}
\description{
This method returns the list of existing jamendo radios
}
\examples{
\donttest{
##Example
 radios <- getRadios(limit=3)
}
}
