% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{getReviewsAlbums}
\alias{getReviewsAlbums}
\title{This method lets you filter and browse album reviews}
\usage{
getReviewsAlbums(
  album_ids,
  limit = 10,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID")
)
}
\arguments{
\item{album_ids}{List of album IDs to seach for}

\item{limit}{Integer to set the number of displayed search results. Maximum value is 200.}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains reviews of the searched albums.
See \url{https://developer.jamendo.com/v3.0/reviews/albums} for more information.
}
\description{
This method lets you filter and browse album reviews
}
\examples{
\donttest{
##Example
 IDs <- c(104336,124067)
 reviews <- getReviewsAlbums(IDs,limit=3)
}
}
