% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{getReviewsAlbumsUser}
\alias{getReviewsAlbumsUser}
\title{This method lets you filter and browse album reviews from a specific user}
\usage{
getReviewsAlbumsUser(user_id, limit = 10,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{user_id}{User ID to seach for}

\item{limit}{Integer to set the number of displayed search results. Maximum value is 200.}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains reviews of the searched albums from a specific user.
See \url{https://developer.jamendo.com/v3.0/reviews/albums} for more information.
}
\description{
This method lets you filter and browse album reviews from a specific user
}
