% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\name{getTracks}
\alias{getTracks}
\title{Search Jamendo database for several tracks}
\usage{
getTracks(track_ids, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{track_ids}{List of track IDs to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about several tracks.
See \url{https://developer.jamendo.com/v3.0/tracks} for more information.
}
\description{
Search Jamendo database for several tracks
}
