% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{getUserProfile}
\alias{getUserProfile}
\title{Search Jamendo database for a user}
\usage{
getUserProfile(user_id = NULL, user_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{user_id}{User ID to seach for}

\item{user_name}{User name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains information about a user.
See \url{https://developer.jamendo.com/v3.0/users} for more information.
}
\description{
Search Jamendo database for a user. You can search for either a user ID or a user name.
}
\examples{
\donttest{
## Example
 userName <- getUserProfile(user_name = "claudod")
 userID <- getUserProfile(user_id = "972174")
}
}
