% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{getUserProfileAlbums}
\alias{getUserProfileAlbums}
\title{Get albums a user added to myalbums}
\usage{
getUserProfileAlbums(
  user_id = NULL,
  user_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID")
)
}
\arguments{
\item{user_id}{User ID to seach for}

\item{user_name}{User name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains albums a user added to myalbums.
See \url{https://developer.jamendo.com/v3.0/users/albums} for more information.
}
\description{
Get albums added to myalbums. You can search for either a user ID or a user name.
}
\examples{
\donttest{
## Example
 albumsFromName <- getUserProfileAlbums(user_name = "claudod")
 albumsFromID <- getUserProfileAlbums(user_id = "972174")
}
}
