% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{getUserProfileArtists}
\alias{getUserProfileArtists}
\title{Get artists a user is a fan of}
\usage{
getUserProfileArtists(user_id = NULL, user_name = NULL,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{user_id}{User ID to seach for}

\item{user_name}{User name to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains artists a user is a fan of.
See \url{https://developer.jamendo.com/v3.0/users/artists} for more information.
}
\description{
Get artists a user is a fan of. You can search for either a user ID or a user name.
}
