% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{getUserProfileTracks}
\alias{getUserProfileTracks}
\title{Get tracks a user has liked, added to favorite or reviewed}
\usage{
getUserProfileTracks(user_id, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{user_id}{User ID to seach for}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"}
}
\value{
Returns a data frame which contains tracks a user has liked, added to favorite or reviewed.
See \url{https://developer.jamendo.com/v3.0/users/tracks} for more information.
}
\description{
Get tracks a user has liked, added to favorite or reviewed
}
\examples{
\donttest{
## Example
 tracks <- getUserProfileTracks("972174")
}
}
