% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parseResponse}
\alias{parseResponse}
\title{Parse response from POST request for setUser functions}
\usage{
parseResponse(res)
}
\arguments{
\item{res}{String from POST request}
}
\value{
Returns string with response message
}
\description{
\code{parse} parses response from POST request for setUser functions.
}
