% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{setUserDislike}
\alias{setUserDislike}
\title{Dislike the track given by Track ID}
\usage{
setUserDislike(token, track_id, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{track_id}{Track ID of rtrack you want to dislike}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"
See \url{https://developer.jamendo.com/v3.0/setuser/dislike} for more information.}
}
\description{
Dislike the track given by Track ID
}
\examples{
\donttest{
## Example
 token <- jamendoOAuth(app_name="xxxx")
 setUserDislike(token, "114069")
}
}
