% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{setUserFan}
\alias{setUserFan}
\title{Become a fan of an artist}
\usage{
setUserFan(token, artist_id, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{artist_id}{Artist ID of artist you want to become a fan of}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"
See \url{https://developer.jamendo.com/v3.0/setuser/fan} for more information.}
}
\description{
Become a fan of an artist
}
