% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{setUserFavorite}
\alias{setUserFavorite}
\title{Add a given track to your preferites}
\usage{
setUserFavorite(token, track_id,
  client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{track_id}{Track ID of track you want to favourite}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"
See \url{https://developer.jamendo.com/v3.0/setuser/favorite} for more information.}
}
\description{
Add a given track to your preferites
}
