% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{setUserLike}
\alias{setUserLike}
\title{Like the track given by Track ID}
\usage{
setUserLike(token, track_id, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{track_id}{Track ID of track you want to like}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"
See \url{https://developer.jamendo.com/v3.0/setuser/like} for more information.}
}
\description{
Like the track given by Track ID
}
