% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usersProfile.R
\name{setUserMyalbum}
\alias{setUserMyalbum}
\title{Add album to your list of favorite albums myalbums}
\usage{
setUserMyalbum(token, album_id, client_id = Sys.getenv("JAMENDO_CLIENT_ID"))
}
\arguments{
\item{token}{Your OAuth access token}

\item{album_id}{Album ID of album you want to add to list myalbums}

\item{client_id}{Defaults to System Environment variable "JAMENDO_CLIENT_ID"
See \url{https://developer.jamendo.com/v3.0/setuser/myalbum} for more information.}
}
\description{
Add album to your list of favorite albums myalbums
}
\examples{
\donttest{
## Example
 token <- jamendoOAuth(app_name="xxxx")
 setUserMyalbum(token, "14866")
}
}
